- Prevent writing a file twice
git-svn-id: svn+ssh://svn.xp-framework.net/home/svn/xpforge/trunk/experiments@11849 d2cacbed-c0f6-0310-851a-9dad52fc623a

+++ arena/zip/io/zip/ZipFileOutputStream.class.php
@@ -73,6 +73,8 @@
      *
      */
     public function close() {
+      if (NULL === $this->data) return;     // Already written
+
       $this->writer->writeFile($this->file, $this->compression, $this->data);
       $this->data= NULL;
     }
